Update map ids to ints

+++ Magic/src/main/java/com/elmakers/mine/bukkit/action/builtin/CycleEntityAction.java
@@ -58,10 +58,10 @@ public class CycleEntityAction extends BaseSpellAction {
                 }
                 int mapId = InventoryUtils.getMapId(frameItem);
                 mapId++;
-                MapView mapView = DeprecatedUtils.getMap((short)mapId);
+                MapView mapView = DeprecatedUtils.getMap(mapId);
                 if (mapView == null) {
                     mapId = 0;
-                    mapView = DeprecatedUtils.getMap((short)mapId);
+                    mapView = DeprecatedUtils.getMap(mapId);
                     if (mapView == null) {
                         return SpellResult.NO_TARGET;
                     }
